Tarefa1_2022li1g093

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

validaObstaculos :: Mapa -> Bool

Exercício 1: A função validaObstaculos, verifica se existem obstáculos em sítios que não devem estar, por exemplo, árvores em rios, troncos em estradas, etc. É utilizada uma função auxiliar para ser mais fácil percorrer a lista de obstáculos dependendo do tipo de terreno, e pelo facto de 2 dos 3 terrenos serem "tipados" como um conjunto de 2 propriendades, o tipo de terreno e a velocidade dos obstáculos

percorreOstaculos :: Terreno -> [Obstaculo] -> Bool

verificaDirecoesRios :: Mapa -> Bool

Exercício 2: A função verificaDirecoesRios , verifica se, todos os rios contíguos tem velociadas para direções diferentes, ou seja, se um rio tem velocidade de 2 o próximo rio contíguo tem que ter velocidade negativa.

verificaDirecoesRios_aux :: [(Terreno, [Obstaculo])] -> Int -> Bool

validaTronco :: Mapa -> Bool

Exercício 3: A função validaTronco , verifica se, para as linhas de Rio, apenas existem troncos com tamanho menor que 5. Caso contrário devolve False invalidando o mapa. Foi necessário usar um acumulador (acc) para contar o número de vezes que o obstáculo tronco ocorre na linha. Foi, mais tarde, pensado como o programa deveria reagir se existi-se uma lista de obstáculos que, devido á regra de todos os objetos que saem dum lado da lista voltam a entrar do outro lado, representasse um tronco com tamanho maior que 5 separado nos cantos da lista, ou seja, a lista: [Tronco, Tronco, Tronco, Tronco, Tronco, Nenhum, Tronco] apresenta um tronco de tamanho 6 que está dividido no final da lista.

validaTronco_aux :: [(Terreno, [Obstaculo])] -> Int -> Bool

validaTroncosSeparados :: Mapa -> Bool

contaTroncosSeguidosEsquerda :: [Obstaculo] -> Int -> Int

contaTroncosSeguidosDireita :: [Obstaculo] -> Int -> Int

validaCarro :: Mapa -> Bool

Exercício 4: A função validaCarro , verifica se, para as linhas de Estrada, apenas existem carros com tamanho menor ou igual a 3. Caso contrário devolve False invalidando o mapa. Foi necessário usar um acumulador (acc) para contar o número de vezes que o obstáculo carro ocorre na linha. Foi, mais tarde, pensado como o programa deveria reagir se existi-se uma lista de obstáculos que, devido á regra de todos os objetos que saem dum lado da lista voltam a entrar do outro lado, representasse um carro com tamanho maior que 3 separado nos cantos da lista, ou seja, a lista: [Carro, Carro, Carro, Nenhum, Carro] apresenta um carro de tamanho 4 que está dividido no final da lista.

validaCarro_aux :: [(Terreno, [Obstaculo])] -> Int -> Bool

validaCarrosSeparados :: Mapa -> Bool

contaCarrosSeguidosEsquerda :: [Obstaculo] -> Int -> Int

contaCarrosSeguidosDireita :: [Obstaculo] -> Int -> Int

existeNenhumMapa :: Mapa -> Bool

Exercício 5: A função existeNenhumMapa , verifica se, cada linha do mapa, com a ajuda da função existeNenhumLinha, tem pelo menos uma célula Nenhum para haver pelo menos um caminho possível para o jogador. Mais tarde será relevante também verificar se as células Nenhum de cada linha estão alinhadas para, realmente haver caminho possível.

existeNenhumlinha :: [Obstaculo] -> Bool

validaComprimentoMapa :: Mapa -> Bool

Exercício 6: A função validaComprimentoMapa , verifica se, o tamanho da lista de obstáculos corresponde á largura indicada no mapa.

validaContigualidade :: Mapa -> Bool

Exercício 7: A função validaContigualidade , verifica se, existem mais de 4 linhas seguidas que são Rios e mais que 5 linhas de estradas e relva.

validaContigualidade_aux :: Mapa -> Int -> Bool

numeroContigualidade_Rio :: Mapa -> Int -> Int

numeroContigualidade_Relva :: Mapa -> Int -> Int

numeroContigualidade_Estrada :: Mapa -> Int -> Int