Tarefa2_2022li1g093

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

A função estendeMapa, vai gerar uma nova linha do mapa de maneira "aleatória", apsar de isto ser verdadeiramente impossível em linguagens de programação, principalmente em Haskell em que o mesmo input numa função é tem sempre o mesmo output. O númro aleatório a ser passado na função estendeMapa deve ser introduzido deste maneira head (randomIntsL (ioConvert (randomizacao)) 1) . É usada uma função auxiliar estedeMapa_aux que tem um caso seja a primeira iteração de gerar uma nova linha é criada essa nova linha, e caso seja necessário preencher-la é o feito a partir da estendeLinha. Para selecionar os terrenos, obstáculos e velocidades aleatórios são usadas as funções escolheTerreno, escolheObs e velocidadeAleatoria respetivamente.

estendeMapa_aux :: Mapa -> Int -> Int -> Mapa

estendeLinha :: Mapa -> Int -> Mapa

escolheTerreno :: [Terreno] -> Int -> Terreno

escolheObs :: [Obstaculo] -> Int -> Obstaculo

velocidadeAleatoria :: Terreno -> Terreno

1 para positivo 0 para negativo, por alguma razão o where vel não funciona, por isso a função fica feia :(

randomizacao :: IO Int

A maneira em que consegui remendar este problema foi usando as funções randomizacao, ioConvert e randomIntsL. A função randomizacao extraí os segundos do sistema usando a função getCurrentTime da lib Data.Time.Clock, depois é passado este valor para um número inteiro, mas esta função devolve um IO Int, agora é possível passar de IO Int para Int mas é necessário usar uma função unsafePerformIO que, como o nome indica, faz uma transformação de IO Int para Int de maneira unsafe, não é recomendado a utilização deste função mas para os objetivos deste trabalho não notei alterações no comportamento do código. Por fim a função randomIntsL foi fornecida pela equipa docente e recebendo uma seed e um tamanho gera uma lista de números aleatórios.

ioConvert :: IO Int -> Int

Função que converte o output da função randomização de IO Int para Int

randomIntsL :: Int -> Int -> [Int]

Função facultada pelos docentes

proximosTerrenosValidos :: Mapa -> [Terreno]

A função @proximosTerrenosValidos recebe um mapa e devolve o tipo de terreno que pode ser inserido numa nova linha em cima do mapa (na cabeça da lista mapa) seguindo as regras estabelecidas na tarefa 1, como a impossibilidade de haver mais do que 5 estradas seguidas, etc. A função auxiliar recebe apenas as primeiras 5 linhas do mapa, uma vez que apenas interessa para pôr de fora os terrenos que ja estão repetidos. São udados acumuladores para os 3 tipos de terrenos para contar o número de linhas desse terreno seguidas.

proximosTerrenosValidos_aux :: [(Terreno, [Obstaculo])] -> Int -> Int -> Int -> [Terreno]

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos recebe um inteiro que representa a largura do mapa e uma linha do mesmo. Para os casos em que o mapa está vazio, e por isso, estamos no início de gerar uma nova linha deve ser possível inserir qualquer tipo de obstaculo, isto feito pela função proximosObstaculosValidos_listaVazia . São utilizadas as funções existeNenhumalinha, contaCarrosSeguidoDireita, contaCarrosSeguidosEsquerda, contaTroncosSeguidosDireita e contaTroncosSeguidosEsquerda definidos na tarefa 1 para verificar se existe pelo menos uma célula Nenhum, e se existe um tronco ou carro maior que 5 ou 3 unidades respetivamente que estejam "divididos" visualmente nos cantos da linha.

proximosObstaculosValidos_listaVazia :: (Terreno, [Obstaculo]) -> [Obstaculo]

verificaComprimentoSeparado_Rio :: [Obstaculo] -> [Obstaculo]

verificaComprimentoSeparado_Estrada :: [Obstaculo] -> [Obstaculo]

proximosObstaculosValidos_Rio :: Int -> [Obstaculo] -> [Obstaculo]

proximosObstaculosValidos_Estrada :: Int -> [Obstaculo] -> [Obstaculo]

proximosObstaculosValidos_Relva :: Int -> [Obstaculo] -> [Obstaculo]