Tarefa3_2022li1g093

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo, usa todas as funções até agora defenidas para animar o jogo dependendo da jogada feita pelo utilizador, foi feita uma divisão de casos a partir de cada movimento, e posteriormente cada movimento implementa a animação para cada linha e ou tipo de terreno. Não interessa aqui se o jogador faz uma jogada que o vai fazer perder o jogo por isso a única coisa a verificar é se o jogador é "bloqueado" por uma árvore, isto é, se a célula para que se pretende mover esteja ocupada por uma árvore, daí a função jogadorBloquiadoArvore, verifica se o jogador deve ser impedido de mover para onde queria. Além disto, na situação em que o jogador se encontra em cima de um tronco, se premanecer parado o jogador deve acompanhar o movimento do tronco e se o jogador se mova quer para a direita quer para a esquerda devem ser somado os vetores de movimento, ou seja, o jogador acompanha o tronco e ainda anda quer para a direita quer para a esquerda.

procuraTerrenoJogador :: Jogo -> Int -> Terreno

animaJogoParado :: Jogo -> Terreno -> Jogo

animaJogoMoveCima :: Jogo -> Terreno -> Jogo

animaJogoMoveBaixo :: Jogo -> Terreno -> Jogo

animaJogoMoveDireita :: Jogo -> Terreno -> Jogo

animaJogoMoveEsquerda :: Jogo -> Terreno -> Jogo

encontraCelulaY :: Mapa -> (Int, Int) -> Int -> Obstaculo

encontraCelulaX :: [Obstaculo] -> (Int, Int) -> Int -> Obstaculo

jogadorBloquiadoArvore :: Jogo -> Jogada -> Bool

moverObstaculos :: Jogo -> Jogo

Exercicio 1: A função moverObstaculos , altera, consoalte a velocidade, a posição dos Troncos e Carros na lista de obstáculos, deixando na posição original um Nenhum. O raciocínio pretendido consiste em guardar a posições dos Troncos ou Carros numa lista de inteiros e somar a cada uma destas posições o número de células que irá "viajar", sendo que se esta soma ultrapassar o tamanho da lista de obstáculos é feito um mod para transformar esse número numa posição válida da lista. Após isto, é feita uma passagem pela lista de obstáculos que passa todas as células a Nenhum, posteriormente esta nova lista é atualizada com a lista de inteiros anterior para colocar o Troncos ou Carros nas devidas posições. A solução inicial envolvia usar a função pré-defenida splitAt para dividir a lista de obstáculos em dois na posição equivalente á velocidade (fazendo ajustes com mod e abs para manter a velocidade "válida") e trocar a ordem. Apresentaram-se problemas ao usar os testes defenidos em (Tarefa3_2022li1g093_Spec.hs), pois no caso da primeira linha do jogo1 ao mover-se os troncos 2 unidade para a direita um deles deveria ficar no início da lista resultante, mas isso não seria possível uma vez que este ja se encontrava no snd do tuplo resultante de splitAt. Isto é: >>> snd(splitAt 2 [Nenhum, Nenhum, Tronco, Nenhum, Tronco, Nehum]) : fst(splitAt 2 [Nenhum, Nenhum, Tronco, Nenhum, Tronco, Nehum]) [Tronco,Nenhum,Tronco,Nenhum,Nehum,Nehum] *quando deveria dar [Tronco, Nenhum, Nenhum, Nenhum, Tronco, Nenhum]

moverObstaculos_aux :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

moverCoisas :: Terreno -> [Obstaculo] -> Int -> [Obstaculo]

passarANenhum :: [Obstaculo] -> [Obstaculo]

posicoesObjetivo :: [Obstaculo] -> Int -> Int -> Int -> [Int]

colocaTroncos :: [Obstaculo] -> [Int] -> Int -> [Obstaculo]

colocaCarros :: [Obstaculo] -> [Int] -> Int -> [Obstaculo]

moveJogador :: Jogo -> Jogada -> Jogo

Exercicio 2: A função moverJogador , move o jogador uma unidade para a direção que for introduzida.

jogadorParadoEmTronco :: Jogo -> Jogo

Exercicio 3: A função jogadorParadoEmTronco , procura se o jogador se encontra em cima de um Tronco e muda as suas coordenadas consoante a velocidade de deslocamento do mesmo. É feita uma procura da prosição do jogador no mapa, primeiro linha a linha, com a função procuraY e depois célula a célula, com procuraX. Depois da descoberta e verificação da posição do jogadore me cima do Tronco, são feitos os cálculos necessários para alterar as coordenadas.

procuraY :: (Int, Int) -> [(Terreno, [Obstaculo])] -> Int -> (Int, Int)

procuraX :: (Int, Int) -> (Terreno, [Obstaculo]) -> Int -> Int -> (Int, Int)

moveJogadorTronco :: (Int, Int) -> Int -> Int -> (Int, Int)

limiteMapa :: Jogo -> Jogada -> Jogo

Exercicio 4: A função limiteMapa , pretende verificar se a jogada dada pelo utilizador não põe o jogador fora do mapa. Para isto é feita uma simples verificação das coordenadas resultantes da simulação do movimento. Se a jogada fôr válida, as coordenadas do jogador são atualizadas. Foi feita uma alteração no output desta função para poder ser usada na realização da tarefa4, assim se a função devolver True é porque as coordenadas do jogador são inválidas.

verificarCoordenadas :: Jogo -> Bool

divideMovers :: Jogo -> Jogo

Fase 2 - Exercicio 1: A função moverCarrosCelula recebe um par de terreno e lista de obstáculos, um inteiro que é a velocidade da dada linha e o par de coordenadas do jogador, esta função apenas move os carros, usando a funcção moverCoisas, uma célula de cada vez. Também é feita uma verificação se o carro ja se encontra "por de cima" do jogador, sendo esta a primeira verificação feita. Ao elaborar esta função deparei-me com o problema em relação a como se iria proseguir quando for para animar uma linha que seja estrada, para isto foi criada a função divideMovers que dado um jogo, em que o jogador se encontra numa estrada, divide as linhas que devem ser animadas "normalmente" a partir da função moverObstaculos_aux e a linha especifica em que o jogador se encontra.

moverCarrosCelula :: (Terreno, [Obstaculo]) -> Int -> (Int, Int) -> (Terreno, [Obstaculo])

moverCarrosCelula_aux :: Terreno -> [Obstaculo] -> Int -> [Obstaculo]

colisao :: [Obstaculo] -> (Int, Int) -> Bool

devolveVelocidade :: Terreno -> Int