Tarefa4_2022li1g093

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

A função jogoTerminou , verifica se o jogador se encontra na água de um rio, na mesma posição de um carro, ou até na mesma posição de uma árvore. Para isto é usada a função encontraLinha que indo linha a linha do mapa procura qual é a linha em que o jogador se encontra. Dependendo do tipo de terreno foram feitas várias funções auxiliares, encontraColuna_Rio, encontraColune_Estrada e encontraColuna_Relva que verificam se o jogador se encontra em alguma posição irregular. Também verifica se o jogador se encontra numa coordenada inválida, i.e. fora do mapa, usando a função verificarCoordenadas da tarefa 3.

encontraLinha :: Mapa -> (Int, Int) -> Int -> Bool

encontraColuna_Rio :: [Obstaculo] -> (Int, Int) -> Int -> Bool

encontraColuna_Estrada :: [Obstaculo] -> (Int, Int) -> Int -> Bool

encontraColuna_Relva :: [Obstaculo] -> (Int, Int) -> Int -> Bool