Tarefa5_2022li1g093

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Int -> Jogo -> Jogo

A função deslizaJogo é a função que irá fazer o efeito de movimento do mapa gerando uma linha nova no topo do mapa usando a função estendeMapa já definida na tarefa 2 e removendo a última linha do mesmo

removeUltima :: Mapa -> Mapa

Função que remove a última linha do mapa

removeUltima_aux :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

Função auxiliar da removeUltima